adding a stitching schema with modifications to acquisition to share classes
